% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeBinary.R
\name{decodeBinary}
\alias{decodeBinary}
\title{Deserialize object}
\usage{
decodeBinary(bin, sep = "-", package = NULL, install_if_not_available = TRUE)
}
\arguments{
\item{bin}{(\code{character(1L)}) Binary string value containing the serialized model.}

\item{sep}{(\code{character(1L)}) Separator used to collapse the binary elements (default is \code{-}).}

\item{package}{(\code{character(1L)}) Package required for object deserialization (default is \code{NULL}).}

\item{install_if_not_available}{(\code{logical(1L)}) Install package if it is not installed (default is \code{TRUE}).}
}
\value{
Deserialized object from \code{bin}
}
\description{
Decode a given string of a serialized object.
}
\examples{
mod = lm(Sepal.Width ~ ., data = iris)
bin = encodeObject(mod)
mod_b = decodeBinary(bin)
all.equal(mod, mod_b)
}
\author{
Daniel S.
}
