% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeModel.R
\name{decodeModel}
\alias{decodeModel}
\title{Deserialize model}
\usage{
decodeModel(bin, sep = "-", package = NULL, install_if_not_available = TRUE)
}
\arguments{
\item{bin}{[character(1L)] Binary string value containing the serialized model.}

\item{sep}{[character(1L)] Separator used to collapse the binary elements.}

\item{package}{[character(1L)] Package required for model predictions.}

\item{install_if_not_available}{[logical(1L)] Install package if it is not installed.}
}
\value{
Model object
}
\description{
Decode given string of serialized model. The deserialized model is then
  assigned to a new variable
}
\examples{
mod = lm(Sepal.Width ~ ., data = iris)
bin = encodeModel(mod)
mod_b = decodeModel(bin)
all.equal(mod, mod_b)
}
\author{
Daniel S.
}
