% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeModel.R
\name{encodeModel}
\alias{encodeModel}
\title{Serialize model}
\usage{
encodeModel(mod, mod_name = NULL, sep = "-", check_serialization = TRUE)
}
\arguments{
\item{mod}{[arbitrary] R object containing a model which is used for predictions on the server side.}

\item{mod_name}{[character(1L)] Name of the model.}

\item{sep}{[character(1L)] Separator used to collapse the binary elements.}

\item{check_serialization}{[logical(1L)] Check if the serialized model can be deserialized
locally.}
}
\value{
Character of length 1 containing the serialized model as string.
}
\description{
This function serializes a given model and creates a character of length 1
  which can be send via DataSHIELD to a server. The string then can be deserialized on the
  server and used to make predictions on the server side.
}
\examples{
mod = lm(Sepal.Width ~ ., data = iris)
bin = encodeModel(mod)
substr(bin, 1, 50)
}
\author{
Daniel S.
}
