% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushModel.R
\name{pushModel}
\alias{pushModel}
\title{Push model to DataSHIELD server}
\usage{
pushModel(
  connections,
  mod,
  sep = "-",
  check_serialization = TRUE,
  package = NULL,
  install_if_not_available = TRUE
)
}
\arguments{
\item{connections}{[DSI::connection] Connection to an OPAL server.}

\item{mod}{[arbitrary] R object containing a model which is used for predictions on the server side.}

\item{sep}{[character(1L)] Separator used to collapse the binary elements.}

\item{check_serialization}{[logical(1L)] Check if the serialized model can be deserialized}

\item{package}{[character(1L)] Package required for model predictions.}

\item{install_if_not_available}{[logical(1L)] Install package if it is not installed.}
}
\description{
This function pushes a serialized model and assigns a server
  variable to it.
}
\author{
Daniel S.
}
