% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMissings.R
\name{removeMissings}
\alias{removeMissings}
\title{Remove missing values from data}
\usage{
removeMissings(symbol)
}
\arguments{
\item{symbol}{(\code{character(1L)}) R object from which missings should be removed.}
}
\value{
Object defined in symbol wrapped by \code{na.omit}.
}
\description{
Remove missing values from data and assign it to new data or old data.
}
\author{
Daniel S.
}
