% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeBinary.R
\name{decodeBinary}
\alias{decodeBinary}
\title{Deserialize object}
\usage{
decodeBinary(bin, package = NULL)
}
\arguments{
\item{bin}{(\code{character(1L)}) Binary string value containing the serialized model.}

\item{package}{(\code{character(1L)}) Package required for object deserialization (default is \code{NULL}).}
}
\value{
Deserialized object from \code{bin}
}
\description{
Decode a given string of a serialized object.
}
\examples{
mod = lm(Sepal.Width ~ ., data = iris)
bin = encodeObject(mod)
mod_b = decodeBinary(bin)
all.equal(mod, mod_b)
}
\author{
Daniel S.
}
