% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMissings.R
\name{removeMissings}
\alias{removeMissings}
\title{Remove missing values from data}
\usage{
removeMissings(connections, symbol)
}
\arguments{
\item{connections}{\link[DSI:connection]{DSI::connection} Connection to an OPAL server.}

\item{symbol}{\link{character(1L)} R object from which missings should be removed.}

\item{new_symbol}{\link{character(1L)} New R object to which the data is written.}
}
\description{
Remove missing values from data and assign it to new data or old data.
}
\author{
Daniel S.
}
